backend/requirements.txt
# Base de Datos
mysql-connector-python==9.1.0

# Scraping y API
requests==2.32.3
beautifulsoup4==4.12.3

# Backend API (Recomendado para conectar con Vue/React después)
fastapi==0.115.0
uvicorn==0.32.0

# Seguridad y Utilidades
python-dotenv==1.0.1
bcrypt==4.2.0

# Generación de PDF
reportlab==4.2.5